+++ tests/unit/video-bridge-log-redaction.test.ts
@@ -26,7 +26,7 @@ import fs from "node:fs";
 import os from "node:os";
 import path from "node:path";
 
-import { redactVideoTranscriptFieldsForLog } from "../../src/lib/guardrails/videoBridgeSnapshotRedaction.ts";
+import { logClientRawRequestRedacted } from "../../src/lib/guardrails/videoBridgeSnapshotRedaction.ts";
 
 const testDataDir = fs.mkdtempSync(path.join(os.tmpdir(), "omni-video-log-redaction-test-"));
 process.env.DATA_DIR = testDataDir;
@@ -275,11 +275,11 @@ test("Scenario A (adversarial review): a message prepended AFTER the guardrail b
 // `transcript`/`audioTranscript` FIELDS on a structured video part, not a flattened
 // description string. Importing the real chatCore.ts here would pull the full
 // request-pipeline dependency graph (executors, providers, combo routing, DB-backed
-// settings, ...) into the test just to reach one guarded ternary a few hundred lines into
-// a 5900+ line handler, for no additional proof beyond what's below — so this exercises
-// the EXACT call-site pattern
-// (`videoBridgeObserved ? redactVideoTranscriptFieldsForLog(body) : body`) against a fake
-// logClientRawRequest. The pure helper itself has its own thorough suite in
+// settings, ...) into the test just to reach one guarded call a few hundred lines into
+// a 5900+ line handler, for no additional proof beyond what's below — so this calls the
+// REAL exported `logClientRawRequestRedacted` (the exact function chatCore.ts's call site
+// invokes, post file-size-refactor) against a fake logClientRawRequest. The pure redaction
+// helper itself has its own thorough suite in
 // tests/unit/guardrails/videoBridgeSnapshotRedaction.test.ts.
 function fakeReqLogger() {
   const calls: unknown[] = [];
@@ -289,21 +289,6 @@ function fakeReqLogger() {
       calls.push(body);
     },
   };
-}
-
-/** Mirrors chatCore.ts's guarded logClientRawRequest call site verbatim. */
-function callLogClientRawRequestAsChatCoreDoes(
-  reqLogger: ReturnType<typeof fakeReqLogger>,
-  clientRawRequest: { endpoint: string; body: unknown; headers: unknown },
-  videoBridgeObserved: boolean
-) {
-  reqLogger.logClientRawRequest(
-    clientRawRequest.endpoint,
-    videoBridgeObserved
-      ? redactVideoTranscriptFieldsForLog(clientRawRequest.body)
-      : clientRawRequest.body,
-    clientRawRequest.headers
-  );
 }
 
 test("surface 2 (raw snapshot): the fake logClientRawRequest receives a redacted snapshot only when videoBridgeObserved is true", () => {
@@ -326,7 +311,7 @@ test("surface 2 (raw snapshot): the fake logClientRawRequest receives a redacted
   const clientRawRequest = { endpoint: "/v1/chat/completions", body: rawBody, headers: {} };
 
   const observedLogger = fakeReqLogger();
-  callLogClientRawRequestAsChatCoreDoes(observedLogger, clientRawRequest, true);
+  logClientRawRequestRedacted(observedLogger, clientRawRequest, true);
   const observedSnapshot = observedLogger.calls[0];
   assert.ok(
     !JSON.stringify(observedSnapshot).includes(SECRET),
@@ -343,10 +328,18 @@ test("surface 2 (raw snapshot): the fake logClientRawRequest receives a redacted
   );
 
   const nonObservedLogger = fakeReqLogger();
-  callLogClientRawRequestAsChatCoreDoes(nonObservedLogger, clientRawRequest, false);
+  logClientRawRequestRedacted(nonObservedLogger, clientRawRequest, false);
   assert.equal(
     nonObservedLogger.calls[0],
     rawBody,
     "the non-observed path must log the exact same object reference — byte-identical, no clone"
+  );
+
+  const skippedLogger = fakeReqLogger();
+  logClientRawRequestRedacted(skippedLogger, null, true);
+  assert.equal(
+    skippedLogger.calls.length,
+    0,
+    "a missing clientRawRequest must not call logClientRawRequest at all (mirrors the old if-guard)"
   );
 });
